Make review changes
