+++ src/comment-counter.test.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-unused-vars */
 const commentCounter = require('./comment-counter.js');
 
 const ul = document.createElement('ul');
@@ -13,21 +12,9 @@ test('0 comment', () => {
   expect(commentCounter(ul)).toBe('<h3>Comments (0)</h3>');
 });
 
-test('0 comment', () => {
-  const li = document.createElement('li');
-  expect(commentCounter(ul)).toBe('<h3>Comments (0)</h3>');
-});
-
 test('1 comment', () => {
   const li = document.createElement('li');
   ul.appendChild(li);
-  expect(commentCounter(ul)).toBe('<h3>Comments (1)</h3>');
-});
-
-test('1 comment', () => {
-  const li1 = document.createElement('li');
-  const li2 = document.createElement('li');
-  ul.appendChild(li1);
   expect(commentCounter(ul)).toBe('<h3>Comments (1)</h3>');
 });
