+++ src/get-movies.js
@@ -1,0 +1,7 @@
+const getMovies = async (movieAPI) => {
+  const response = await fetch(movieAPI);
+  const data = response.json();
+  return data;
+};
+
+module.exports = getMovies;
